compressed gif
The gif in the README was very long and taking a while to load. An unsuspecting reader of the README might even get the impression that there was no gif there to see, since they would have finished reading the whole thing before it loaded. To fix this issue I have compressed the gif down from 13.7 MB to 4.3 MB