shrunk gif
The earlier gif was dauntingly large. Now it is an acceptable size (shrunk it from W: 750px, H: 1334px to W: 400px, H: 711px). This should make it easier to fit the whole gif on a screen at once. It also shrinks its file size down from 4.12 MiB to 1.43 MiB.